backend/requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-environ>=0.11
celery[redis]>=5.3
redis>=5.0
Pillow>=10.0
drf-spectacular>=0.26
django-cors-headers>=4.3
django-filter>=23.0
reportlab>=4.0
openpyxl>=3.1
gunicorn>=21.0
